+++ app/src/main/java/com/example/the_movie_db_test_task/ui/screens/MainScreen.kt
@@ -15,9 +15,10 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import com.example.the_movie_db_test_task.R
-import com.example.the_movie_db_test_task.data.model.discaver.Movie
+import com.example.the_movie_db_test_task.data.api.discaver.Movie
 import com.example.the_movie_db_test_task.ui.items.TabBar
 import com.example.the_movie_db_test_task.ui.theme.Purple500
+import com.example.the_movie_db_test_task.utils.Constants
 
 @Composable
 fun MainScreen() {
@@ -42,7 +43,6 @@ fun MainScreen() {
                 contentScale = ContentScale.Crop
             )
         }
-
         TabScreen()
     }
 }
@@ -51,8 +51,7 @@ fun MainScreen() {
 @Composable
 fun TabScreen() {
     var tabIndex by remember { mutableStateOf(0) }
-    val tabs = listOf("Films", "Favorite")
-    var movieDB = mutableStateOf<MutableList<Movie>>(mutableListOf())
+    val tabs = listOf(Constants.films, Constants.favorite)
 
     Column(modifier = Modifier.fillMaxWidth()) {
         TabRow(selectedTabIndex = tabIndex) {
