+++ app/src/main/java/com/example/the_movie_db_test_task/ui/screens/films/FilmsScreen.kt
@@ -11,14 +11,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.Constraints
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.the_movie_db_test_task.data.model.discaver.Movie
+import com.example.the_movie_db_test_task.data.api.discaver.Movie
 import com.example.the_movie_db_test_task.ui.items.DataItem
 import com.example.the_movie_db_test_task.ui.items.MovieCardItem
-import com.example.the_movie_db_test_task.ui.screens.favorites.SharedViewModel
 import com.example.the_movie_db_test_task.utils.Constants
+import com.example.the_movie_db_test_task.utils.convertData
 import com.google.accompanist.swiperefresh.SwipeRefresh
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
 import java.text.DateFormat
@@ -27,59 +26,87 @@ import java.util.*
 
 @Composable
 fun FilmsScreen(viewModel: SharedViewModel = hiltViewModel()) {
-    val outputFormat: DateFormat = SimpleDateFormat("MMM yyyy")
-    var caunter = 1
-    val movie by remember { viewModel.movie }
-    val endReached by remember { viewModel.endReached }
+    val movies by remember { viewModel.movies }
+    val moviesDB by remember { viewModel.moviesDB }
     val isLoading by remember { viewModel.isLoading }
     val isError by remember { viewModel.isError }
-    val swipeRefreshState = rememberSwipeRefreshState(isRefreshing = isLoading)
-    var date: Date? = null
+
+
     Column(
         modifier = Modifier.fillMaxSize(),
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center
     ) {
         when {
-            isLoading -> {
-                CircularProgressIndicator(
-                    color = Color.Blue,
-
-                    )
-            }
-            isError -> {
-                Text(text = "ERROR")
-            }
+            isError -> Text(text = Constants.error)
             else -> {
-                SwipeRefresh(state = swipeRefreshState, onRefresh = { viewModel.getMovies() }) {
-                    LazyColumn(contentPadding = PaddingValues(16.dp),
-                        modifier = Modifier.fillMaxSize()
-                    ) {
-                        items(movie) { item ->
-                            if (item.id == movie.last().id) {
-                                viewModel.getMovies()
-                            }
-                            if (item.releaseDate.isNotEmpty()) {
-                                if (date == null) {
-                                    date = convertData(item.releaseDate)
-                                    DataItem(data = outputFormat.format(date))
-                                } else {
-                                    if (date!! > convertData(item.releaseDate)) {
-                                        date = convertData(item.releaseDate)
-                                        DataItem(data = outputFormat.format(date))
-                                    }
-                                }
-                            }
-                            MovieCardItem(
-                                movie = item,
-                                onClickSecondBut = {},
-                                textFirstBat = if (viewModel.movieDB.value.any { it.id == item.id }) {
-                                    Constants.unlike
-                                } else {
-                                    Constants.like
-                                },
-                                onClickFirstBut = { viewModel.postFavorite(item) })
+                if (isLoading && movies.isEmpty()) { CircularProgressIndicator(color = Color.Blue)
+                } else {
+                        MovieList(
+                            movies = movies,
+                            isLoading = isLoading,
+                            moviesDB = moviesDB,
+                            viewModel = viewModel
+                        )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun MovieList(
+    movies: List<Movie>,
+    isLoading: Boolean,
+    moviesDB: List<Movie>,
+    viewModel: SharedViewModel
+) {
+    var date: Date? = remember { null }
+    val outputFormat: DateFormat = SimpleDateFormat("MMM yyyy")
+    val swipeRefreshState = rememberSwipeRefreshState(isRefreshing = isLoading)
+    SwipeRefresh(state = swipeRefreshState, onRefresh = { viewModel.getMovies() }) {
+
+        LazyColumn(
+            contentPadding = PaddingValues(16.dp),
+            modifier = Modifier.fillMaxSize()
+        ) {
+            items(
+                movies,
+                Movie::id
+            ) { item ->
+                if (item.id == movies.last().id) {
+                    viewModel.getMovies()
+                }
+                if (item.releaseDate.isNotEmpty()) {
+                    date?.let {
+                        if (it > convertData(item.releaseDate)) {
+                            date = convertData(item.releaseDate)
+                            DataItem(data = outputFormat.format(date))
                         }
+                    }?.run {
+                        date = convertData(item.releaseDate)
+                        DataItem(data = outputFormat.format(date))
+                    }
+                }
+                MovieCardItem(
+                    movie = item,
+                    onClickSecondBut = {},
+                    textFirstBat = if (moviesDB.any { it.id == item.id }) {
+                        Constants.unlike
+                    } else {
+                        Constants.like
+                    },
+                    onClickFirstBut = { viewModel.postFavorite(item) })
+            }
+            item {
+                if (isLoading) {
+                    Column(
+                        modifier = Modifier
+                            .fillMaxWidth(),
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                        verticalArrangement = Arrangement.Center,
+                    ) {
+                        CircularProgressIndicator(color = Color.Black)
                     }
                 }
             }
@@ -87,17 +114,6 @@ fun FilmsScreen(viewModel: SharedViewModel = hiltViewModel()) {
     }
 }
 
-fun convertData(inputData: String): Date? {
-    return when {
-        inputData.isNotEmpty() -> {
-            val inputFormat: DateFormat = SimpleDateFormat("yyyy-MM-dd")
-            val outputFormat: DateFormat = SimpleDateFormat("MMM yyyy")
-            val date: Date? = inputFormat.parse(inputData)
-            val stringData = date?.let { outputFormat.format(it) }
-            stringData?.let { outputFormat.parse(it) }
-        }
-        else -> {
-            null
-        }
-    }
-}
+
+
+
