+++ app/src/main/java/com/example/the_movie_db_test_task/utils/Constants.kt
@@ -1,7 +1,6 @@
 package com.example.the_movie_db_test_task.utils
 
 
-
 //Ключ API
 //a9cfabbe527a6111ab2f5c1d5fa8f3f3
 
@@ -13,14 +12,13 @@ package com.example.the_movie_db_test_task.utils
 object Constants {
     const val BASE_URL = "https://api.themoviedb.org/3/"
     const val API_TOKEN = "Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJhOWNmYWJiZTUyN2E2MTExYWIyZjVjMWQ1ZmE4ZjNmMyIsInN1YiI6IjY1MWJjNTg5YzUwYWQyMDEwYmZmNGMxYSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.gcIUonMVa7Q2r92rgvbo-RyvigSnP_VO4LMQ9KJvTPM"
-
-    //title for params
+    const val error = "ERROR"
+    const val accountId = 20523672
     const val like = "Like"
     const val unlike = "Unlike"
     const val share = "share"
     const val films = "Films"
     const val favorite = "Favorite"
     const val remove = "Remove"
-
 
 }
